Fix more windows warnings

See merge request GNOME/glib!2303

+++ glib/gwin32.c
@@ -534,7 +534,7 @@ g_win32_check_windows_version (const gint major,
   HMODULE hmodule;
 #endif
   /* We Only Support Checking for XP or later */
-  g_return_val_if_fail (major >= 5 && (major <=6 || major == 10), FALSE);
+  g_return_val_if_fail (major >= 5 && (major <= 6 || major == 10), FALSE);
   g_return_val_if_fail ((major >= 5 && minor >= 1) || major >= 6, FALSE);
 
   /* Check for Service Pack Version >= 0 */
@@ -553,14 +553,14 @@ g_win32_check_windows_version (const gint major,
   RtlGetVersion (&osverinfo);
 
   /* check the OS and Service Pack Versions */
-  if (osverinfo.dwMajorVersion > major)
+  if (osverinfo.dwMajorVersion > (DWORD) major)
     is_ver_checked = TRUE;
-  else if (osverinfo.dwMajorVersion == major)
+  else if (osverinfo.dwMajorVersion == (DWORD) major)
     {
-      if (osverinfo.dwMinorVersion > minor)
+      if (osverinfo.dwMinorVersion > (DWORD) minor)
         is_ver_checked = TRUE;
-      else if (osverinfo.dwMinorVersion == minor)
-        if (osverinfo.wServicePackMajor >= spver)
+      else if (osverinfo.dwMinorVersion == (DWORD) minor)
+        if (osverinfo.wServicePackMajor >= (DWORD) spver)
           is_ver_checked = TRUE;
     }
